relocating UpdateFailureProcessor file to bot folder
