+++ src/main/java/com/learning/urlshortener/bot/UrlShortenerBot.java
@@ -14,7 +14,6 @@ import org.telegram.telegrambots.extensions.bots.commandbot.commands.IBotCommand
 import org.telegram.telegrambots.meta.api.objects.Update;
 
 import com.learning.urlshortener.bot.commands.noncommand.NonCommandUpdateHandler;
-import com.learning.urlshortener.bot.commands.noncommand.UpdateFailureProcessor;
 import com.learning.urlshortener.bot.utils.MessageUtils;
 import com.learning.urlshortener.bot.utils.TgIncomingUpdateLogger;
