+++ src/main/java/com/learning/urlshortener/bot/testbot/UrlShortenerTestBot.java
@@ -11,7 +11,7 @@ import org.telegram.telegrambots.meta.api.methods.BotApiMethod;
 
 import com.learning.urlshortener.bot.UrlShortenerBot;
 import com.learning.urlshortener.bot.commands.noncommand.NonCommandUpdateHandler;
-import com.learning.urlshortener.bot.commands.noncommand.UpdateFailureProcessor;
+import com.learning.urlshortener.bot.UpdateFailureProcessor;
 import com.learning.urlshortener.bot.utils.MessageUtils;
 import com.learning.urlshortener.bot.utils.TgIncomingUpdateLogger;
